content/html/content/test/file_iframe_sandbox_b_if3.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for Bug 341604</title>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<script>
  function ok(result, message) {
    window.parent.postMessage({ok: result, desc: message}, "*");
  }

  function testXHR() {
  var xhr = new XMLHttpRequest();

  xhr.open("GET", "file_iframe_sandbox_b_if1.html");

  xhr.onreadystatechange = function (oEvent) {
    var result = false;
    if (xhr.readyState == 4) {
      if (xhr.status == 0) {
        result = true;
      }
      ok(result, "XHR should be blocked in an iframe sandboxed WITHOUT 'allow-same-origin'");
    }
  }

  xhr.send(null);
}

  function doStuff() {
    try {
      window.parent.ok(false, "documents sandboxed without 'allow-same-origin' should NOT be able to access their parent");
    } catch (error) {
      ok(true, "documents sandboxed without 'allow-same-origin' should NOT be able to access their parent");
    }

    // should NOT be able to access document.cookie
    try {
      var foo = document.cookie;
    } catch(error) {
      ok(true, "a document sandboxed without allow-same-origin should NOT be able to access document.cookie");
    }

    // should NOT be able to access localStorage
    try {
      var foo = window.localStorage;
    } catch(error) {
      ok(true, "a document sandboxed without allow-same-origin should NOT be able to access localStorage");
    }

    // should NOT be able to access sessionStorage
    try {
      var foo = window.sessionStorage;
    } catch(error) {
      ok(true, "a document sandboxed without allow-same-origin should NOT be able to access sessionStorage");
    }

  testXHR();
  }
</script>
<body onLoad="doStuff()">
  I am sandboxed but with "allow-scripts"
</body>
</html>
